add error handlers and custom error classes

+++ src/tags/service.py
@@ -1,4 +1,3 @@
-
 from fastapi import status
 from fastapi.exceptions import HTTPException
 from sqlmodel import desc, select
@@ -8,6 +7,7 @@ from src.books.service import BookService
 from src.db.models import Tag
 
 from .schemas import TagAddModel, TagCreateModel
+from src.errors import BookNotFound, TagNotFound, TagAlreadyExists
 
 book_service = BookService()
 
@@ -36,12 +36,10 @@ class TagService:
         book = await book_service.get_book(book_uid=book_uid, session=session)
 
         if not book:
-            raise HTTPException(status_code=404, detail="Book not found")
+            raise BookNotFound()
 
         for tag_item in tag_data.tags:
-            result = await session.exec(
-                select(Tag).where(Tag.name == tag_item.name)
-            )
+            result = await session.exec(select(Tag).where(Tag.name == tag_item.name))
 
             tag = result.one_or_none()
             if not tag:
@@ -52,8 +50,6 @@ class TagService:
         await session.commit()
         await session.refresh(book)
         return book
-
-
 
     async def get_tag_by_uid(self, tag_uid: str, session: AsyncSession):
         """Get tag by uid"""
@@ -74,10 +70,7 @@ class TagService:
         tag = result.first()
 
         if tag:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN, detail="Tag exists"
-            )
-
+            raise TagAlreadyExists()
         new_tag = Tag(name=tag_data.name)
 
         session.add(new_tag)
@@ -93,6 +86,9 @@ class TagService:
 
         tag = await self.get_tag_by_uid(tag_uid, session)
 
+        if not tag:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+
         update_data_dict = tag_update_data.model_dump()
 
         for k, v in update_data_dict.items():
@@ -104,7 +100,6 @@ class TagService:
 
         return tag
 
-
     async def delete_tag(self, tag_uid: str, session: AsyncSession):
         """Delete a tag"""
 
@@ -114,8 +109,9 @@ class TagService:
 
         tag = result.first()
 
+        if not tag:
+            raise TagNotFound()
+
         await session.delete(tag)
 
         await session.commit()
-
-
